+++ CHANGELOG.md
@@ -4,4 +4,5 @@ CHANGELOG
 master
 ------
 
+* Add a button to optimize image on demand
 * New LiipMonitor check to verify number of compression run during the last month through TinyPNG API